Added user_id on clients

+++ src/tests/client.rs
@@ -4,20 +4,20 @@ use serde_json;
 use uuid::Uuid;
 use rocket_contrib::json::Json;
 
-use crate::{get_rocket_instance, models::client::{Client, NewClient}};
+use crate::{get_rocket_instance, models::client::{Client, NewClient}, models::user::{User, NewUser}};
 
 #[test]
 fn create() {
     let client = rocket::local::Client::new(get_rocket_instance()).expect("valid rocket instance");
 
-    let (response_client, response_client_expected) = create_client();
+    let (response_client, response_client_expected, user_id) = create_client();
 
     assert_eq!(response_client.name, response_client_expected.name);
     assert_eq!(response_client.email, response_client_expected.email);
     assert_eq!(response_client.phone_number, response_client_expected.phone_number);
     assert_eq!(response_client.address, response_client_expected.address);
 
-    client.delete(format!("/clients/{}",response_client.id))
+    client.delete(format!("/clients/{}/{}",user_id,response_client.id))
         .header(ContentType::JSON)
         .dispatch();
 }
@@ -26,9 +26,9 @@ fn create() {
 fn show() {
     let client = rocket::local::Client::new(get_rocket_instance()).expect("valid rocket instance");
 
-    let (response_client_create, response_client_expected) = create_client();
+    let (response_client_create, response_client_expected,user_id) = create_client();
 
-    let mut response = client.get(format!("/clients/{}",response_client_create.id))
+    let mut response = client.get(format!("/clients/{}/{}",user_id,response_client_create.id))
     .header(ContentType::JSON)
     .dispatch();
 
@@ -40,7 +40,7 @@ fn show() {
     assert_eq!(response_client.phone_number, response_client_expected.phone_number);
     assert_eq!(response_client.address, response_client_expected.address);
 
-    client.delete(format!("/clients/{}",response_client.id))
+    client.delete(format!("/clients/{}/{}",user_id,response_client.id))
         .header(ContentType::JSON)
         .dispatch();
 }
@@ -49,19 +49,34 @@ fn show() {
 fn delete() {
     let client = rocket::local::Client::new(get_rocket_instance()).expect("valid rocket instance");
 
-    let (response_client, response_client_expected) = create_client();
+    let (response_client, response_client_expected,user_id) = create_client();
 
-    let response = client.delete(format!("/clients/{}",response_client.id))
+    let response = client.delete(format!("/clients/{}/{}",user_id,response_client.id))
         .header(ContentType::JSON)
         .dispatch();
 
     assert_eq!(response.status(), Status::Ok);
 }
 
-fn create_client() -> (Client, Client){
+fn create_client() -> (Client, Client, Uuid){
     let client = rocket::local::Client::new(get_rocket_instance()).expect("valid rocket instance");
 
+    let users = NewUser{
+        name: "Quick",
+        email: "henrique.fquick@gmail.com",
+        password: "123",
+        share_photos: None,
+        darkmode: Some(false)
+    };
+
+    let mut response_create = client.post("/users")
+        .header(ContentType::JSON)
+        .body(serde_json::to_string(&users).unwrap())
+        .dispatch();
+    let response_user_create: User = serde_json::from_str(response_create.body_string().unwrap().as_str()).unwrap();
+
     let clients = NewClient{
+        user_id: response_user_create.id,
         name: "Quick",
         email: "henrique.fquick@gmail.com",
         phone_number: "31998180608",
@@ -75,13 +90,14 @@ fn create_client() -> (Client, Client){
 
     let response_client_expected = Client{
             id: response_client_create.id,
+            user_id: response_user_create.id,
             name: String::from("Quick"),
             email: Some(String::from("henrique.fquick@gmail.com")),
             phone_number: Some(String::from("31998180608")),
             address: Some(String::from("Rua teste"))
     };
 
-    let mut response = client.get(format!("/clients/{}",response_client_create.id))
+    let mut response = client.get(format!("/clients/{}/{}",response_user_create.id,response_client_create.id))
         .header(ContentType::JSON)
         .body(serde_json::to_string(&clients).unwrap())
         .dispatch();
@@ -90,5 +106,5 @@ fn create_client() -> (Client, Client){
 
     let response_client: Client = serde_json::from_str(response.body_string().unwrap().as_str()).unwrap();
 
-    return (response_client, response_client_expected);
+    return (response_client, response_client_expected,response_user_create.id);
 }
